pyproject.toml
[project]
name = "comeback-agency"
version = "0.1.0"
description = "Book Management System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "python-jose>=3.5.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
